requirements.txt
Flask==1.1.2
mysql-connector
requests